+++ src/main/java/com/x1/frans/supplier/query/service/SupplierQueryService.java
@@ -5,5 +5,5 @@ import com.x1.frans.supplier.query.dto.SupplierDetailDTO;
 public interface SupplierQueryService {
     String findLatestCodeByCodePrefix(String codePrefix);
 
-    SupplierDetailDTO getSupplierDetail(Long supplierId);
+    SupplierDetailDTO getSupplierDetail(int supplierId);
 }
